% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Data set for construction of selection index}
\format{
A data frame with 689 rows and 7 columns:
\describe{
  \item{sire}{Sire of the cows}
  \item{farm}{Farm from which the data were collected}
  \item{soc}{The season of calving of a cow}
  \item{poc}{The period of calving of a cow}
  \item{tmy}{Total lactation milk yield in Kg}
  \item{py}{The peak yield in Kg}
  \item{afc}{The age at first calving}
  \item{fatyield}{The average fat yield}
}
}
\usage{data(data, package="selINDRIx")}
\description{
This dataset contains information used for constructing a selection index.
}
\keyword{datasets}
