% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_si.R
\name{random_si}
\alias{random_si}
\title{Title Construction of selection index}
\usage{
random_si(data_input, traits, economic_values)
}
\arguments{
\item{traits}{The traits for which index values are to be estimated}

\item{economic_values}{The relative economic values}

\item{data}{A data frame containing the fixed, random and traits}

\item{random}{The random effects}
}
\value{
Results of selection index
}
\description{
Title Construction of selection index
}
\examples{
results <- mixed_si(data = data, traits = traits,
random = random, economic_values = economic_values)

}
